I am trying to extract bounding box for each character. craft_utils.py -> get_poly_core give error for multiple bounding boxes.

+++ detection/craft_text_detector/craft_text_detector/utils/get_detection_boxes_core.py
@@ -111,7 +111,7 @@ def make_box(segmap, only_characters=False):
             box.append(prop.bbox)
         # make it compatible
         box = np.array(box)
-        compatible_box = np.zeros((box.shape[0], 4, 2))
+        compatible_box = np.zeros((box.shape[0], 4, 2), dtype=np.float32)
 
         box_coord_left_bottom = np.array([box[:, 1], box[:, 2]]).transpose()
         compatible_box[:, 0, :] = box_coord_left_bottom
